+++ module/src/calculators/calculateAperture.test.ts
